Moving getLayerSurf to a full <long int> from size_t   This allows the surface to be a complement

+++ Model/essBuild/ButterflyModerator.cxx
@@ -242,7 +242,7 @@ ButterflyModerator::getCommonSurf(const size_t) const
 }
 
 int
-ButterflyModerator::getLayerSurf(const size_t,const size_t) const
+ButterflyModerator::getLayerSurf(const size_t,const long int) const
   /*!
     Only components have reference values
   */
