+++ Model/essBuildInc/EdgeWater.h
@@ -71,7 +71,7 @@ class EdgeWater :
 
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
-  virtual int getLayerSurf(const size_t,const size_t) const;
+  virtual int getLayerSurf(const size_t,const long int) const;
 
   void createAll(Simulation&,const attachSystem::FixedComp&,
 		 const std::string&,const std::string&);
