+++ Model/photonInc/CylContainer.h
@@ -3,7 +3,7 @@
  
  * File:   photonInc/CylContainer.h
  *
- * Copyright (c) 2004-2015 by Stuart Ansell
+ * Copyright (c) 2004-2016 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -78,7 +78,7 @@ class CylContainer :
   virtual int getMainBody() const { return cylIndex+1; }
 
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
-  virtual int getLayerSurf(const size_t,const size_t) const;
+  virtual int getLayerSurf(const size_t,const long int) const;
   virtual int getCommonSurf(const size_t) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
