+++ Model/t1UpgradeInc/CylPreSimple.h
@@ -80,7 +80,7 @@ class CylPreSimple : public attachSystem::ContainedComp,
 		 const attachSystem::FixedComp&);
 
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
-  virtual int getLayerSurf(const size_t,const size_t) const;
+  virtual int getLayerSurf(const size_t,const long int) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
